requirements.txt
opencv-python>=4.5.0
cohere>=5.0.0
pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.64.0
pandas>=1.5.0
matplotlib>=3.5.0
numpy>=1.21.0
pytest>=7.0.0
elevenlabs>=2.15.0
flask>=2.3.0
flask-cors>=4.0.0
